Refactor: EyeTutorialManager+Configure extension으로 분리
