+++ APPRO/APPRO/Views/Tutorial/Eye/EyeTutorialManager.swift
@@ -155,66 +155,3 @@ extension EyeTutorialManager {
     }
     
 }
-
-// MARK: - Entity Configuration & Load Methods
-
-private extension EyeTutorialManager {
-    
-    func loadEntity(entityType: EntityType) async throws -> Entity {
-        return try await Entity(named: entityType.rawValue, in: realityKitContentBundle)
-    }
-    
-    func configureEyeEntity(entity: Entity) {
-        setClosureComponent(entity: entity, distance: .eyes)
-    }
-    
-    func configurePatchEntity(eyeEntity: Entity) {
-        guard let patchEntity = eyeEntity.findEntity(named: "patch") else {
-            dump("configurePatchEntity failed: patch not found in eyeEntity")
-            return
-        }
-        patchEntity.components.set(HoverEffectComponent(.highlight(.default)))
-    }
-    
-    func configureChickenEntity(entity: Entity) {
-        entity.name = chickenEntityName
-        entity.setPosition(.init(x: 2, y: 0, z: 0), relativeTo: eyeEntity)
-        entity.components.set(OpacityComponent(opacity: 0.0))
-    }
-    
-    func setClosureComponent(
-        entity: Entity,
-        distance: Float,
-        upward: Float = 0,
-        forwardDirection: Entity.ForwardDirection = .negativeZ
-    ) {
-        let closureComponent = ClosureComponent { [weak self] deltaTime in
-            guard let currentTransform = self?.headTracker.originFromDeviceTransform() else { return }
-            
-            let currentTranslation = currentTransform.translation()
-            let targetPosition = currentTranslation - distance * currentTransform.forward()
-            let ratio = Float(pow(0.96, deltaTime / (16 * 1E-3)))
-            var newPosition = ratio * entity.position(relativeTo: nil) + (1 - ratio) * targetPosition
-            newPosition.y += upward
-            entity.look(at: currentTranslation, from: newPosition, relativeTo: nil, forward: forwardDirection)
-        }
-        entity.components.set(closureComponent)
-    }
-    
-}
-
-private enum EntityType: String {
-    
-    case eyes = "Eye/eyes_with_patch.usd"
-    case ring = "Eye/eye_ring.usd"
-    case chicken = "Eye/chicken.usd"
-    case attachment
-    
-}
-
-private extension Float {
-    
-    static let eyes = Float(2.0)
-    static let attachment = Float(2.05)
-    
-}
